requirements.txt
PyYAML
confluent-kafka
sickle
PyMongo
requests